bluetooth: host: Allow concurrent advertising with multiple ids
The HCI specification creates additional complexity to allow this
configuration:
 - When a connection gets established, we need to know which
   identity the HCI_LE_Connection_Complete event corresponds to.
 - The identity is a property of the advertising set.
   Therefore we need the advertising handle.
 - The advertising handle is part of the
   HCI_LE_Advertising_Set_Terminated event and is not part of
   the HCI_LE_Connection_Complete event. Therefore
   the information of both events needs to be combined.

As the caching requires more resources, it is only enabled if the
application requires multiple advertising sets and multiple
identities.

By spec the LE_Connection_Complete comes first. Therefore we cache
this event until the identity is available.

After the controller has forwarded the HCI_LE_Connection_Complete
event, it is responsible for forwarding the corresponding
HCI_LE_Advertising_Set_Terminated event afterwards. Otherwise the
host will not be able to determine the identities of the new
established connections. See Bluetooth spec errata 14871 for more
details about this controller requirement.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>